+++ firebase-crashlytics/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Unreleased
+
+
+# 19.0.3
 * [changed] Update the internal file system to handle long file names.
+
+
+## Kotlin
+The Kotlin extensions library transitively includes the updated
+`firebase-crashlytics` library. The Kotlin extensions library has no additional
+updates.
 
 # 19.0.2
 * [changed] Changing caught exception type to fail safely on any exception type.